util: Split into multiple files
