+++ vita3k/util/include/util/string_utils.h
@@ -17,7 +17,11 @@
 
 #pragma once
 
+#include <algorithm>
+#include <codecvt>
 #include <cstdint>
+#include <locale>
+#include <sstream>
 #include <string>
 #include <vector>
