+++ vita3k/util/src/byte.cpp
@@ -1,0 +1,52 @@
+// Vita3K emulator project
+// Copyright (C) 2023 Vita3K team
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#include <util/bytes.h>
+
+template <>
+uint16_t byte_swap(uint16_t val) {
+    return (val >> 8) | (val << 8);
+}
+
+template <>
+uint32_t byte_swap(uint32_t val) {
+    //        AA              BB00                      CC0000                       DD000000
+    return (val >> 24) | ((val >> 8) & 0xFF00) | ((val << 8) & 0xFF0000) | ((val << 24) & 0xFF000000);
+}
+
+template <>
+uint64_t byte_swap(uint64_t val) {
+    val = ((val << 8) & 0xFF00FF00FF00FF00ULL) | ((val >> 8) & 0x00FF00FF00FF00FFULL);
+    val = ((val << 16) & 0xFFFF0000FFFF0000ULL) | ((val >> 16) & 0x0000FFFF0000FFFFULL);
+
+    return (val << 32) | (val >> 32);
+}
+
+template <>
+int16_t byte_swap(int16_t val) {
+    return byte_swap(static_cast<uint16_t>(val));
+}
+
+template <>
+int32_t byte_swap(int32_t val) {
+    return byte_swap(static_cast<uint32_t>(val));
+}
+
+template <>
+int64_t byte_swap(int64_t val) {
+    return byte_swap(static_cast<uint64_t>(val));
+}
