--- vita3k/util/src/util.cpp
@@ -1,883 +1,0 @@
-// Vita3K emulator project
-// Copyright (C) 2023 Vita3K team
-//
-// This program is free software; you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation; either version 2 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License along
-// with this program; if not, write to the Free Software Foundation, Inc.,
-// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
-
-#include <string_view>
-#include <util/arm.h>
-#include <util/bytes.h>
-#include <util/log.h>
-#include <util/net_utils.h>
-#include <util/string_utils.h>
-#include <util/vector_utils.h>
-
-#include <spdlog/sinks/basic_file_sink.h>
-#include <spdlog/sinks/msvc_sink.h>
-#include <spdlog/sinks/stdout_color_sinks.h>
-
-#ifdef WIN32
-#define WIN32_LEAN_AND_MEAN
-#include <Windows.h>
-#else
-#include <signal.h>
-#include <sys/mman.h>
-#include <unistd.h>
-#endif
-
-#ifdef WIN32
-#include <winsock2.h>
-#else
-#include <fcntl.h>
-#endif
-
-#include <algorithm>
-#include <bitset>
-#include <codecvt> // std::codecvt_utf8
-#include <iostream>
-#include <locale> // std::wstring_convert
-#include <map>
-#include <memory>
-#include <set>
-#include <sstream>
-#include <string>
-
-#include <memory>
-#include <stdexcept>
-
-namespace logging {
-
-static const fs::path &LOG_FILE_NAME = "vita3k.log";
-static const char *LOG_PATTERN = "%^[%H:%M:%S.%e] |%L| [%!]: %v%$";
-std::vector<spdlog::sink_ptr> sinks;
-
-void register_log_exception_handler();
-
-void flush() {
-    spdlog::details::registry::instance().flush_all();
-}
-
-ExitCode init(const Root &root_paths, bool use_stdout) {
-    sinks.clear();
-    if (use_stdout)
-        sinks.push_back(std::make_shared<spdlog::sinks::stdout_color_sink_mt>());
-
-    if (add_sink(root_paths.get_log_path() / LOG_FILE_NAME) != Success)
-        return InitConfigFailed;
-
-    spdlog::set_error_handler([](const std::string &msg) {
-        std::cerr << "spdlog error: " << msg << std::endl;
-        assert(0);
-    });
-
-#ifdef WIN32
-    // set console codepage to UTF-8
-    SetConsoleOutputCP(65001);
-    SetConsoleTitle("Vita3K PSVita Emulator");
-#endif
-
-    register_log_exception_handler();
-
-    static std::terminate_handler old_terminate = nullptr;
-    old_terminate = std::set_terminate([]() {
-        try {
-            std::rethrow_exception(std::current_exception());
-        } catch (const std::exception &e) {
-            LOG_CRITICAL("Unhandled C++ exception. {}", e.what());
-        } catch (...) {
-            LOG_CRITICAL("Unhandled C++ exception. UNKNOWN");
-        }
-        flush();
-        if (old_terminate)
-            old_terminate();
-    });
-    return Success;
-}
-
-void set_level(spdlog::level::level_enum log_level) {
-    spdlog::set_level(log_level);
-}
-
-ExitCode add_sink(const fs::path &log_path) {
-    try {
-#ifdef WIN32
-        sinks.push_back(std::make_shared<spdlog::sinks::basic_file_sink_mt>(log_path.generic_path().wstring(), true));
-#else
-        sinks.push_back(std::make_shared<spdlog::sinks::basic_file_sink_mt>(log_path.generic_path().string(), true));
-#endif
-    } catch (const spdlog::spdlog_ex &ex) {
-        std::cerr << "File log initialization failed: " << ex.what() << std::endl;
-        return InitConfigFailed;
-    }
-
-#ifdef _MSC_VER
-    if (sinks.size() == 2) { // spdlog is being initialized
-        sinks.push_back(std::make_shared<spdlog::sinks::msvc_sink_mt>());
-    }
-#endif
-
-    spdlog::set_default_logger(std::make_shared<spdlog::logger>("vita3k logger", begin(sinks), end(sinks)));
-    spdlog::set_pattern(LOG_PATTERN);
-    return Success;
-}
-
-// log exceptions and flush log file on exceptions
-#ifdef WIN32
-static LONG WINAPI exception_handler(PEXCEPTION_POINTERS pExp) noexcept {
-    const unsigned ec = pExp->ExceptionRecord->ExceptionCode;
-    switch (ec) {
-    case EXCEPTION_ACCESS_VIOLATION:
-        LOG_CRITICAL("Exception EXCEPTION_ACCESS_VIOLATION ({}). ", log_hex(ec));
-        switch (pExp->ExceptionRecord->ExceptionInformation[0]) {
-        case 0:
-            LOG_CRITICAL("Read violation at address {}.", log_hex(pExp->ExceptionRecord->ExceptionInformation[1]));
-            break;
-        case 1:
-            LOG_CRITICAL("Write violation at address {}.", log_hex(pExp->ExceptionRecord->ExceptionInformation[1]));
-            break;
-        case 8:
-            LOG_CRITICAL("DEP violation at address {}.", log_hex(pExp->ExceptionRecord->ExceptionInformation[1]));
-            break;
-        default:
-            break;
-        }
-        break;
-    case EXCEPTION_ARRAY_BOUNDS_EXCEEDED:
-        LOG_CRITICAL("Exception EXCEPTION_ARRAY_BOUNDS_EXCEEDED ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_DATATYPE_MISALIGNMENT:
-        LOG_CRITICAL("Exception EXCEPTION_DATATYPE_MISALIGNMENT ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_FLT_DIVIDE_BY_ZERO:
-        LOG_CRITICAL("Exception EXCEPTION_FLT_DIVIDE_BY_ZERO ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_ILLEGAL_INSTRUCTION:
-        LOG_CRITICAL("Exception EXCEPTION_ILLEGAL_INSTRUCTION ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_IN_PAGE_ERROR:
-        LOG_CRITICAL("Exception EXCEPTION_IN_PAGE_ERROR ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_INT_DIVIDE_BY_ZERO:
-        LOG_CRITICAL("Exception EXCEPTION_INT_DIVIDE_BY_ZERO ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_PRIV_INSTRUCTION:
-        LOG_CRITICAL("Exception EXCEPTION_PRIV_INSTRUCTION ({}). ", log_hex(ec));
-        break;
-    case EXCEPTION_STACK_OVERFLOW:
-        LOG_CRITICAL("Exception EXCEPTION_STACK_OVERFLOW ({}). ", log_hex(ec));
-        break;
-    default:
-        return EXCEPTION_CONTINUE_SEARCH;
-    }
-    flush();
-    return EXCEPTION_CONTINUE_SEARCH;
-}
-
-void register_log_exception_handler() {
-    if (!AddVectoredExceptionHandler(0, exception_handler)) {
-        LOG_CRITICAL("Failed to register an exception handler");
-    }
-}
-
-#else
-void register_log_exception_handler() {}
-#endif
-} // namespace logging
-
-#ifdef TRACY_ENABLE
-#include <util/tracy_module_utils.h>
-
-namespace tracy_module_utils {
-
-constexpr int max_modules = 64; // If not enough increase to 64*n
-typedef std::bitset<max_modules> tracy_module_flags;
-typedef std::vector<std::string> tracy_module_names;
-
-tracy_module_names tracy_available_advanced_profiling_modules{};
-tracy_module_flags tracy_advanced_profiling_modules{};
-
-tracy_module_helper::tracy_module_helper(const char *module_name) {
-    mod_id = tracy_available_advanced_profiling_modules.size();
-    tracy_available_advanced_profiling_modules.emplace_back(module_name);
-    if (mod_id >= max_modules) {
-        LOG_ERROR_ONCE("Too many tracy modules. Increase max_modules const");
-    }
-}
-
-bool is_tracy_active(tracy_module_helper module) {
-    return tracy_advanced_profiling_modules.test(module.mod_id);
-}
-
-bool is_tracy_active(const std::string &module) {
-    int module_index = vector_utils::find_index(tracy_available_advanced_profiling_modules, module);
-    if (module_index >= 0)
-        return tracy_advanced_profiling_modules.test(module_index);
-    else
-        return false;
-}
-
-void set_tracy_active(const std::string &module, bool value) {
-    int module_index = vector_utils::find_index(tracy_available_advanced_profiling_modules, module);
-    if (module_index >= 0) {
-        tracy_advanced_profiling_modules.set(module_index, value);
-    }
-}
-
-std::vector<std::string> get_available_module_names() {
-    std::vector<std::string> names = tracy_available_advanced_profiling_modules;
-    std::sort(names.begin(), names.end());
-    return names;
-}
-
-void load_from(const std::vector<std::string> &active_modules_str) {
-    tracy_advanced_profiling_modules.reset();
-    for (auto &module : active_modules_str)
-        set_tracy_active(module, true);
-}
-
-void cleanup(std::vector<std::string> &active_modules_str) {
-    // remove if not found in tracy_available_advanced_profiling_modules
-    std::erase_if(active_modules_str, [](const std::string &module) {
-        return std::find(tracy_available_advanced_profiling_modules.begin(), tracy_available_advanced_profiling_modules.end(), module) == tracy_available_advanced_profiling_modules.end();
-    });
-}
-
-} // namespace tracy_module_utils
-#endif
-
-namespace string_utils {
-
-std::vector<std::string> split_string(const std::string &str, char delimiter) {
-    std::stringstream str_stream(str);
-    std::string segment;
-    std::vector<std::string> seglist;
-
-    const size_t num_segments = std::count_if(str.begin(), str.end(), [&](char c) {
-        return c == delimiter;
-    }) + (str.empty() ? 1 : 0);
-
-    seglist.reserve(num_segments);
-
-    while (std::getline(str_stream, segment, delimiter)) {
-        seglist.push_back(segment);
-    }
-    return seglist;
-}
-
-std::wstring utf_to_wide(const std::string &str) {
-    std::wstring_convert<std::codecvt_utf8<wchar_t>> myconv;
-    return myconv.from_bytes(str);
-}
-
-std::string wide_to_utf(const std::wstring &str) {
-    std::wstring_convert<std::codecvt_utf8<wchar_t>> myconv;
-    return myconv.to_bytes(str);
-}
-
-std::string remove_special_chars(std::string str) {
-    for (char &c : str) {
-        switch (c) {
-        case '/':
-        case '\\':
-        case ':':
-        case '?':
-        case '"':
-        case '<':
-        case '>':
-        case '|':
-        case '*':
-            c = ' ';
-            break;
-        default:
-            continue;
-        }
-    }
-    return str;
-}
-
-// Based on: https://stackoverflow.com/a/23135441
-// Search and replace "in" with "out" in the given string
-void replace(std::string &str, const std::string &in, const std::string &out) {
-    size_t pos = 0;
-    while ((pos = str.find(in, pos)) != std::string::npos) {
-        str.replace(pos, in.length(), out);
-        pos += out.length();
-    }
-}
-
-std::basic_string<uint8_t> string_to_byte_array(std::string string) {
-    std::basic_string<uint8_t> hex_bytes;
-
-    for (size_t i = 0; i < string.length(); i += 2) {
-        uint16_t byte;
-        std::string nextbyte = string.substr(i, 2);
-        std::istringstream(nextbyte) >> std::hex >> byte;
-        hex_bytes.push_back(static_cast<uint8_t>(byte));
-    }
-    return hex_bytes;
-}
-
-#ifdef WIN32
-std::string utf16_to_utf8(const std::u16string &str) {
-    std::wstring_convert<std::codecvt_utf8_utf16<int16_t>, int16_t> myconv;
-    auto p = reinterpret_cast<const int16_t *>(str.data());
-    return myconv.to_bytes(p, p + str.size());
-}
-
-std::u16string utf8_to_utf16(const std::string &str) {
-    static_assert(sizeof(std::wstring::value_type) == sizeof(std::u16string::value_type),
-        "std::wstring and std::u16string are expected to have the same character size");
-
-    std::wstring_convert<std::codecvt_utf8_utf16<int16_t>, int16_t> myconv;
-    auto p = reinterpret_cast<const char *>(str.data());
-    auto a = myconv.from_bytes(p, p + std::strlen(p));
-    return std::u16string(a.begin(), a.end());
-}
-#else
-std::string utf16_to_utf8(const std::u16string &str) {
-    std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> myconv;
-    return myconv.to_bytes(str);
-}
-
-std::u16string utf8_to_utf16(const std::string &str) {
-    std::wstring_convert<std::codecvt_utf8_utf16<char16_t>, char16_t> myconv;
-    return myconv.from_bytes(str);
-}
-#endif
-
-std::string toupper(const std::string &s) {
-    std::string r = s;
-    std::transform(r.begin(), r.end(), r.begin(),
-        [](unsigned char c) { return std::toupper(c); });
-    return r;
-}
-
-std::string tolower(const std::string &s) {
-    std::string r = s;
-    std::transform(r.begin(), r.end(), r.begin(),
-        [](unsigned char c) { return std::tolower(c); });
-    return r;
-}
-
-int stoi_def(const std::string &str, int default_value, const char *name) {
-    try {
-        return std::stoi(str);
-    } catch (std::invalid_argument &e) {
-        LOG_ERROR("Invalid {}: \"{}\"", name, str);
-    } catch (std::out_of_range &e) {
-        LOG_ERROR("Out of range {}: \"{}\"", name, str);
-    }
-    return default_value;
-}
-
-} // namespace string_utils
-
-namespace net_utils {
-
-// 0 is ok, negative is bad
-SceHttpErrorCode parse_url(std::string url, parsedUrl &out) {
-    out.scheme = url.substr(0, url.find(':'));
-
-    if (out.scheme != "http" && out.scheme != "https")
-        return SCE_HTTP_ERROR_UNKNOWN_SCHEME;
-
-    // Check if URL is opaque, if it is opaque then its invalid
-    // http://
-    //+    012
-    //      ^^ Check these 2
-    char url1Slash = *(url.c_str() + (out.scheme.length() + 1)); // get uint8 value
-    char url2Slash = *(url.c_str() + (out.scheme.length() + 2)); // get next uint8 value
-    // Compare the 2 characters that are supposed to be slahses
-    if (url1Slash != '/' || url2Slash != '/') {
-        out.invalid = true;
-        return (SceHttpErrorCode)0;
-    }
-
-    auto end_scheme_pos = url.find(':');
-    auto path_pos = url.find('/', end_scheme_pos + 3);
-    auto has_path = path_pos != std::string::npos;
-
-    auto full_wo_scheme = url.substr(end_scheme_pos + 3);
-
-    if (has_path) {
-        // username:password@lttstore.com:727/wysi/cookie.php?pog=gers#extremeexploit
-
-        {
-            path_pos = std::string(full_wo_scheme).find('/');
-            auto full_no_scheme_path = full_wo_scheme.substr(0, path_pos);
-            // full_no_scheme_path = username:password@lttstore.com:727
-            auto c = full_no_scheme_path.find('@');
-            if (c != std::string::npos) { // we have credentials
-                auto credentials = full_no_scheme_path.substr(0, c);
-                // credentials = username:password
-                auto semicolon_pos = credentials.find(':');
-
-                if (semicolon_pos != std::string::npos) { // we have password?
-                    auto username = credentials.substr(0, semicolon_pos);
-                    auto password = credentials.substr(semicolon_pos + 1);
-
-                    if (username.length() >= SCE_HTTP_USERNAME_MAX_SIZE)
-                        return SCE_HTTP_ERROR_OUT_OF_SIZE;
-                    if (password.length() >= SCE_HTTP_USERNAME_MAX_SIZE)
-                        return SCE_HTTP_ERROR_OUT_OF_SIZE;
-
-                    out.username = username;
-                    out.password = password;
-                } else { // we don't have password
-                    out.username = credentials;
-                }
-            } else { // no credentials
-                // lttstore.com:727
-                auto semicolon_pos = full_no_scheme_path.find(':');
-
-                if (semicolon_pos != std::string::npos) { // we have port
-                    auto hostname = full_no_scheme_path.substr(0, semicolon_pos);
-                    auto port = full_no_scheme_path.substr(semicolon_pos + 1);
-
-                    out.hostname = hostname;
-                    out.port = port;
-                } else { // no port
-                    out.hostname = full_no_scheme_path;
-                }
-            }
-        }
-        {
-            auto path_pos = std::string(full_wo_scheme).find('/');
-            auto full_no_scheme_hostname = full_wo_scheme.substr(path_pos);
-            // /wysi/cookie.php?pog=gers#extremeexploit
-
-            auto query_pos = full_no_scheme_hostname.find('?');
-            if (query_pos != std::string::npos) { // we have query
-                auto path = full_no_scheme_hostname.substr(0, query_pos);
-                out.path = path;
-
-                auto query_frag = full_no_scheme_hostname.substr(query_pos);
-
-                auto frag_pos = query_frag.find('#');
-                if (frag_pos != std::string::npos) {
-                    // query and fragment
-
-                    auto query = query_frag.substr(0, frag_pos);
-                    auto fragment = query_frag.substr(frag_pos);
-
-                    out.query = query;
-                    out.fragment = fragment;
-                } else { // no fragment
-                    out.query = query_frag;
-                }
-
-            } else { // no query, dunno about fragment
-                auto frag_pos = full_no_scheme_hostname.find('#');
-
-                if (frag_pos != std::string::npos) { // we have fragment
-                    auto path = full_no_scheme_hostname.substr(0, frag_pos);
-                    auto fragment = full_no_scheme_hostname.substr(frag_pos);
-
-                    out.path = path;
-                    out.fragment = fragment;
-                } else { // no fragment, only path
-                    out.path = full_no_scheme_hostname;
-                }
-            }
-        }
-    } else {
-        // username:password@lttstore.com:727
-        auto c = full_wo_scheme.find('@');
-        if (c != std::string::npos) { // we have credentials
-            auto credentials = full_wo_scheme.substr(0, c);
-            // credentials = username:password
-            auto semicolon_pos = credentials.find(':');
-
-            if (semicolon_pos != std::string::npos) { // we have password?
-                auto username = credentials.substr(0, semicolon_pos);
-                auto password = credentials.substr(semicolon_pos + 1);
-
-                if (username.length() >= SCE_HTTP_USERNAME_MAX_SIZE)
-                    return SCE_HTTP_ERROR_OUT_OF_SIZE;
-                if (password.length() >= SCE_HTTP_USERNAME_MAX_SIZE)
-                    return SCE_HTTP_ERROR_OUT_OF_SIZE;
-
-                out.username = username;
-                out.password = password;
-            } else { // we don't have password
-                out.username = credentials;
-            }
-        } else { // no credentials
-            // lttstore.com:727
-            auto semicolon_pos = full_wo_scheme.find(':');
-
-            if (semicolon_pos != std::string::npos) { // we have port
-                auto hostname = full_wo_scheme.substr(0, semicolon_pos);
-                auto port = full_wo_scheme.substr(semicolon_pos + 1);
-
-                out.hostname = hostname;
-                out.port = port;
-            } else { // no port
-                out.hostname = full_wo_scheme;
-            }
-        }
-    }
-
-    return (SceHttpErrorCode)0;
-}
-
-int char_method_to_int(const char *method) {
-    if (strcmp(method, "GET") == 0) {
-        return SCE_HTTP_METHOD_GET;
-    } else if (strcmp(method, "POST") == 0) {
-        return SCE_HTTP_METHOD_POST;
-    } else if (strcmp(method, "HEAD") == 0) {
-        return SCE_HTTP_METHOD_HEAD;
-    } else if (strcmp(method, "OPTIONS") == 0) {
-        return SCE_HTTP_METHOD_OPTIONS;
-    } else if (strcmp(method, "PUT") == 0) {
-        return SCE_HTTP_METHOD_PUT;
-    } else if (strcmp(method, "DELETE") == 0) {
-        return SCE_HTTP_METHOD_DELETE;
-    } else if (strcmp(method, "TRACE") == 0) {
-        return SCE_HTTP_METHOD_TRACE;
-    } else if (strcmp(method, "CONNECT") == 0) {
-        return SCE_HTTP_METHOD_CONNECT;
-    } else {
-        return -1;
-    }
-}
-
-const char *int_method_to_char(const int n) {
-    switch (n) {
-    case 0: return "GET";
-    case 1: return "POST";
-    case 2: return "HEAD";
-    case 3: return "OPTIONS";
-    case 4: return "PUT";
-    case 5: return "DELETE";
-    case 6: return "TRACE";
-    case 7: return "CONNECT";
-
-    default:
-        return "INVALID";
-        break;
-    }
-}
-
-std::string constructHeaders(HeadersMapType &headers) {
-    std::string headersString;
-    for (auto head : headers) {
-        headersString.append(head.first);
-        headersString.append(": ");
-        headersString.append(head.second);
-        headersString.append("\r\n");
-    }
-
-    return headersString;
-}
-
-bool parseStatusLine(std::string line, std::string &httpVer, int &statusCode, std::string &reason) {
-    auto lineClean = line.substr(0, line.find("\r\n"));
-
-    // do this check just in case the server is drunk or retarded, would be nice to do more checks with some regex
-    if (!lineClean.starts_with("HTTP/"))
-        return false; // what
-
-    const auto firstSpace = lineClean.find(" ");
-    if (firstSpace == std::string::npos)
-        return false;
-
-    const std::string fullHttpVerStr = lineClean.substr(0, firstSpace);
-    const std::string httpVerStr = fullHttpVerStr.substr(strlen("HTTP/"));
-
-    if (!std::isdigit(httpVerStr[0]))
-        return false;
-
-    if (lineClean.length() < fullHttpVerStr.length() + strlen(" XXX"))
-        return false; // the rest of the line is less than 3 characters in length, what the fuck happened also abort
-
-    const auto codeAndReason = lineClean.substr(firstSpace + 1);
-    const auto statusCodeStr = codeAndReason.substr(0, 3);
-    if (!std::isdigit(statusCodeStr[0]) || !std::isdigit(statusCodeStr[1]) || !std::isdigit(statusCodeStr[2]))
-        return false; // status code contains non digit characters, abort
-
-    const int statusCodeInt = std::stoi(statusCodeStr);
-
-    std::string reasonStr = "";
-    bool hasReason = codeAndReason.find(" ") != std::string::npos;
-    if (hasReason) // standard says that reasons CAN be empty, we have to take this edge case into account
-        reasonStr = codeAndReason.substr(4);
-
-    httpVer = httpVerStr;
-    statusCode = statusCodeInt;
-    reason = reasonStr;
-
-    return true;
-}
-
-/*
-    CANNOT have ANYTHING after the last \r\n or \r\n\r\n else it will be treated as a header
-*/
-bool parseHeaders(std::string &headersRaw, HeadersMapType &headersOut) {
-    char *ptr;
-    ptr = strtok(headersRaw.data(), "\r\n");
-    // use while loop to check ptr is not null
-    while (ptr != NULL) {
-        auto line = std::string_view(ptr);
-
-        if (line.find(':') == std::string::npos)
-            return false; // separator is missing, the header is invalid
-
-        auto name = line.substr(0, line.find(':'));
-        int valueStart = name.length() + 1;
-        if (line.find(": "))
-            // Theres a space between semicolon and value, trim it
-            valueStart++;
-
-        auto value = line.substr(valueStart);
-
-        headersOut.insert({ std::string(name), std::string(value) });
-        ptr = strtok(NULL, "\r\n");
-    }
-    return true;
-}
-
-bool parseResponse(std::string res, SceRequestResponse &reqres) {
-    auto statusLine = res.substr(0, res.find("\r\n"));
-    if (!parseStatusLine(statusLine, reqres.httpVer, reqres.statusCode, reqres.reasonPhrase))
-        return false;
-
-    auto headersRaw = res.substr(res.find("\r\n") + strlen("\r\n"), res.find("\r\n\r\n"));
-
-    if (!parseHeaders(headersRaw, reqres.headers))
-        return false;
-
-    bool hasContLen = false;
-    int contLenVal = 0;
-
-    auto contLenIt = reqres.headers.find("Content-Length");
-    if (contLenIt == reqres.headers.end()) {
-        reqres.contentLength = 0;
-    } else {
-        reqres.contentLength = std::stoi(contLenIt->second);
-    }
-
-    return true;
-}
-
-bool socketSetBlocking(int sockfd, bool blocking) {
-#ifdef WIN32
-    ioctlsocket(sockfd, FIONBIO, (u_long *)&blocking);
-#else
-    if (blocking) { // Blocking
-        int flags = fcntl(sockfd, F_GETFL); // Get flags
-        fcntl(sockfd, F_SETFL, flags & ~O_NONBLOCK); // Set NONBLOCK flag off
-    } else { // Non blocking
-        int flags = fcntl(sockfd, F_GETFL); // Get flags
-        fcntl(sockfd, F_SETFL, flags | O_NONBLOCK); // Set NONBLOCK flag on
-    }
-#endif
-    return true;
-}
-
-} // namespace net_utils
-
-template <>
-std::uint16_t byte_swap(std::uint16_t val) {
-    return (val >> 8) | (val << 8);
-}
-
-template <>
-std::uint32_t byte_swap(std::uint32_t val) {
-    //        AA              BB00                      CC0000                       DD000000
-    return (val >> 24) | ((val >> 8) & 0xFF00) | ((val << 8) & 0xFF0000) | ((val << 24) & 0xFF000000);
-}
-
-template <>
-std::uint64_t byte_swap(std::uint64_t val) {
-    val = ((val << 8) & 0xFF00FF00FF00FF00ULL) | ((val >> 8) & 0x00FF00FF00FF00FFULL);
-    val = ((val << 16) & 0xFFFF0000FFFF0000ULL) | ((val >> 16) & 0x0000FFFF0000FFFFULL);
-
-    return (val << 32) | (val >> 32);
-}
-
-template <>
-std::int16_t byte_swap(std::int16_t val) {
-    return byte_swap(static_cast<std::uint16_t>(val));
-}
-
-template <>
-std::int32_t byte_swap(std::int32_t val) {
-    return byte_swap(static_cast<std::uint32_t>(val));
-}
-
-template <>
-std::int64_t byte_swap(std::int64_t val) {
-    return byte_swap(static_cast<std::uint64_t>(val));
-}
-
-/*
-float32 to float16 conversion
-we can have 3 cases
-1 program compiled with AVX2 or F16C flags  - use fast conversion
-2 program compiled without AVX2 or F16C flags:
-2a we can include and use F16C intrinsic without AVX2 or F16C flags - autodetect and use fast or basic conversion depends of runtime cpu
-2b we can't include and use F16C intrinsic - use basic conversion
-msvc allow to include any intrinsic independent of architecture flags, other compilers disallow this
-*/
-#if (defined(__AVX__) && defined(__F16C__)) || defined(__AVX2__) || (defined(_MSC_VER) && !defined(__clang__))
-#include <immintrin.h>
-void float_to_half_AVX_F16C(const float *src, std::uint16_t *dest, const int total) {
-    int left = total;
-
-    while (left >= 8) {
-        __m256 float_vector = _mm256_loadu_ps(src);
-        __m128i half_vector = _mm256_cvtps_ph(float_vector, 0);
-        _mm_storeu_si128((__m128i *)dest, half_vector);
-
-        left -= 8;
-        src += 8;
-        dest += 8;
-    }
-
-    if (left > 0) {
-        alignas(32) float data[8] = { 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f };
-        std::copy_n(src, left, data);
-        __m256 float_vector = _mm256_load_ps(data);
-        __m128i half_vector = _mm256_cvtps_ph(float_vector, 0);
-        _mm_store_si128(reinterpret_cast<__m128i *>(data), half_vector);
-        std::copy_n(reinterpret_cast<uint16_t *>(data), left, dest);
-    }
-}
-#endif
-
-#if (defined(__AVX__) && defined(__F16C__)) || defined(__AVX2__)
-// forced use AVX+F16C instruction set
-// AVX2 checked intentionally cause MSVC does not have __F16C__ macros
-// and checking AVX is not enough for some CPU architectures (Intel Sandy bridge)
-void float_to_half(const float *src, std::uint16_t *dest, const int total) {
-    float_to_half_AVX_F16C(src, dest, total);
-}
-#elif defined(__aarch64__)
-#include <arm_neon.h>
-void float_to_half(const float *src, std::uint16_t *dest, const int total) {
-    int left = total;
-    while (left >= 4) {
-        float32x4_t floatx4 = vld1q_f32(src);
-        float16x4_t halfx4 = vcvt_f16_f32(floatx4);
-        vst1_f16(reinterpret_cast<float16_t *>(dest), halfx4);
-        left -= 4;
-        src += 4;
-        dest += 4;
-    }
-
-    if (left > 0) {
-        float data[4] = { 0.0, 0.0, 0.0, 0.0 };
-        std::copy_n(src, left, data);
-        float32x4_t floatx4 = vld1q_f32(data);
-        float16x4_t halfx4 = vcvt_f16_f32(floatx4);
-        vst1_f16(reinterpret_cast<float16_t *>(data), halfx4);
-        std::copy_n(reinterpret_cast<uint16_t *>(data), left, dest);
-    }
-}
-#else
-#include <util/float_to_half.h>
-void float_to_half_basic(const float *src, std::uint16_t *dest, const int total) {
-    for (int i = 0; i < total; i++) {
-        dest[i] = util::encode_flt16(src[i]);
-    }
-}
-#if (defined(_MSC_VER) && !defined(__clang__))
-// check and use AVX+F16C instruction set if possible
-
-// use function variable as imitation of self-modifying code.
-// on first use we check processor features and set appropriate realisation, later we immediately use appropriate realisation
-void float_to_half_init(const float *src, std::uint16_t *dest, const int total);
-
-void (*float_to_half_var)(const float *src, std::uint16_t *dest, const int total) = float_to_half_init;
-
-#include <util/instrset_detect.h>
-void float_to_half_init(const float *src, std::uint16_t *dest, const int total) {
-    if (util::instrset::hasF16C()) {
-        float_to_half_var = float_to_half_AVX_F16C;
-        LOG_INFO("AVX+F16C instruction set is supported. Using fast f32 to f16 conversion");
-    } else {
-        float_to_half_var = float_to_half_basic;
-        LOG_INFO("AVX+F16C instruction set is not supported. Using basic f32 to f16 conversion");
-    }
-    (*float_to_half_var)(src, dest, total);
-}
-
-void float_to_half(const float *src, std::uint16_t *dest, const int total) {
-    (*float_to_half_var)(src, dest, total);
-}
-#else
-void float_to_half(const float *src, std::uint16_t *dest, const int total) {
-    float_to_half_basic(src, dest, total);
-}
-#endif
-#endif
-
-// Encode code taken from https://github.com/yifanlu/UVLoader/blob/master/resolve.c
-
-uint32_t encode_arm_inst(uint8_t type, uint32_t immed, uint16_t reg) {
-    switch (type) {
-    case INSTRUCTION_MOVW:
-        // 1110 0011 0000 XXXX YYYY XXXXXXXXXXXX
-        // where X is the immediate and Y is the register
-        // Upper bits == 0xE30
-        return ((uint32_t)0xE30 << 20) | ((uint32_t)(immed & 0xF000) << 4) | (immed & 0xFFF) | (reg << 12);
-    case INSTRUCTION_MOVT:
-        // 1110 0011 0100 XXXX YYYY XXXXXXXXXXXX
-        // where X is the immediate and Y is the register
-        // Upper bits == 0xE34
-        return ((uint32_t)0xE34 << 20) | ((uint32_t)(immed & 0xF000) << 4) | (immed & 0xFFF) | (reg << 12);
-    case INSTRUCTION_SYSCALL:
-        // Syscall does not have any immediate value, the number should
-        // already be in R12
-        return (uint32_t)0xEF000000;
-    case INSTRUCTION_BRANCH:
-        // 1110 0001 0010 111111111111 0001 YYYY
-        // BX Rn has 0xE12FFF1 as top bytes
-        return ((uint32_t)0xE12FFF1 << 4) | reg;
-    case INSTRUCTION_BLX:
-        return ((uint32_t)0x7D << 25) | ((immed & 0x80000000) >> 8) | (immed & 0x7fffff);
-    case INSTRUCTION_UNKNOWN:
-    default:
-        return 0;
-    }
-}
-
-inline uint32_t encode_thumb_blx(uint32_t immed) {
-    const uint32_t S = (immed & 0x80000000) >> 31;
-    const uint32_t I1 = (immed & 0x800000) >> 22;
-    const uint32_t I2 = (immed & 0x400000) >> 21;
-    const uint32_t immhi = (immed & 0x3ff000) >> 11;
-    const uint32_t immlo = (immed & 0xffc) >> 2;
-    const uint32_t J1 = ~I1 ^ S;
-    const uint32_t J2 = ~I2 ^ S;
-    return ((uint32_t)0x1E << 27) | (S << 26) | (immhi << 16) | ((uint32_t)0x3 << 14) | (J1 << 13) | (J2 << 11) | (immlo << 1);
-}
-
-uint32_t encode_thumb_inst(uint8_t type, uint32_t immed, uint16_t reg) {
-    switch (type) {
-    case INSTRUCTION_MOVW:
-        return ((uint32_t)0x1E << 27) | ((uint32_t)(immed & 0x800) << 15) | ((uint32_t)0x24 << 20) | ((immed & 0xf000) << 4) | ((immed & 0x700) << 4) | (reg << 8) | (immed & 0xff);
-    case INSTRUCTION_MOVT:
-        return ((uint32_t)0x1E << 27) | ((uint32_t)(immed & 0x800) << 15) | ((uint32_t)0x2C << 20) | ((immed & 0xf000) << 4) | ((immed & 0x700) << 4) | (reg << 8) | (immed & 0xff);
-    case INSTRUCTION_BRANCH:
-        return ((((uint32_t)0x8E << 7) | (reg << 3)) << 16) | (uint32_t)0xBF00;
-    case INSTRUCTION_BLX:
-        return encode_thumb_blx(immed);
-    case INSTRUCTION_UNKNOWN:
-    default:
-        return 0;
-    }
-}
